:8ball: Proper test setup.
This commit rearranges the library structure. The only hardware
dependent library, the processor library, is substituted with a mock
during unit testing by the platformIO library dependency finder.
This way the actual code does not know about the untit testing files and
the unit test does not know anything about the hardware dependencies.