🐛 Reverts MMIO back to pointers.
It is nice to refer to the MMIO as references, but since I am using
the MMIO type definitions of ST which are C structs, they do not have
default copy and move operators. By using pointers, these can be
implicitly defined by the compiler, which gives all MMIO objects the
default copy and move operators. This can be obviously lead to problems.
But as long as you know what you are doing, the benefits outweigh the
risks, in my opinion.